Where in the world is this IP address?

Where in the world is this IP address?

crm_locate(ip, database = carmen::ip2location_lite_db1, cache = TRUE)

Arguments

ip
A vector of IP addresses. Either character (e.g. "130.191.166.222") or numeric (e.g. 2193598174).

database
A database of IP address locations from IP2Location. Either a data.frame or the path to a .csv file.

cache
If TRUE, the data in database will be cached for this session, improving the speed of subsequent lookups.

Source

This site or product includes IP2Location LITE data available from https://lite.ip2location.com.

Value

A data.frame with a row of location information for each element of ip.

See also

Examples

crm_locate("130.191.166.222")

#> # A tibble: 1 x 2
#>   country_code country_name            
#>   <chr>        <chr>                   
#> 1 US           United States of America

crm_locate(c("130.191.166.222", "190.115.7.226", "201.248.64.43"))

#> # A tibble: 3 x 2
#>   country_code country_name                      
#>   <chr>        <chr>                             
#> 1 US           United States of America          
#> 2 GT           Guatemala                         
#> 3 VE           Venezuela (Bolivarian Republic of)

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

tibble(
  id   = seq_len(100),
  name = generator::r_full_names(100),
  ip   = generator::r_ipv4_addresses(100)
) %>%
  mutate(crm_locate(ip))

#> # A tibble: 100 x 5
#>       id name             ip              country_code country_name            
#>    <int> <chr>            <chr>           <chr>        <chr>                   
#>  1     1 Toby Lubowitz    46.52.197.150   RU           Russian Federation      
#>  2     2 Damian Keeling   39.123.203.16   KR           Korea (Republic of)     
#>  3     3 Paris O'Conner   151.176.97.209  DE           Germany                 
#>  4     4 Jenifer Stracke  7.56.69.228     US           United States of America
#>  5     5 Belkis Hermiston 143.82.226.2    US           United States of America
#>  6     6 Sachiko Corwin   142.145.105.13  CA           Canada                  
#>  7     7 Noah Will        223.211.225.229 CN           China                   
#>  8     8 Ian Pfannerstill 35.152.200.65   ES           Spain                   
#>  9     9 Ruben Herzog     162.22.115.53   US           United States of America
#> 10    10 Clora Hilpert    22.89.145.10    US           United States of America
#> # … with 90 more rows